Separated English, Greek and Greek/English support.
