+++ rapp_speech_detection_sphinx4/src/rapp_speech_detection_sphinx4/__init__.py
@@ -16,6 +16,7 @@
 
 from english_support import EnglishSupport
 from greek_support import GreekSupport
+from greek_english_support import GreekEnglishSupport
 from language_support import LanguageSupport
 from limited_vocabulary_creator import LimitedVocabularyCreator
 from rapp_exceptions import RappError
